pull types out from cennznet-primitives into types.rs
- types are in types.rs
- changed crate imports accordingly
