+++ primitives/src/lib.rs
@@ -14,61 +14,10 @@
 // You should have received a copy of the GNU General Public License
 // along with Substrate.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Low-level types used by CENNZnet node.
+//! Common types and traits used by CENNZnet node.
 
 #![warn(missing_docs)]
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod traits;
-
-use sr_primitives::{
-	generic,
-	traits::{BlakeTwo256, Verify},
-	AnySignature, DoughnutV0, OpaqueExtrinsic,
-};
-
-/// An index to a block.
-pub type BlockNumber = u32;
-
-/// Alias to 512-bit hash when used in the context of a transaction signature on the chain.
-pub type Signature = AnySignature;
-
-/// Some way of identifying an account on the chain. We intentionally make it equivalent
-/// to the public key of our transaction signing scheme.
-pub type AccountId = <Signature as Verify>::Signer;
-
-/// The type for looking up accounts. We don't expect more than 4 billion of them, but you
-/// never know...
-pub type AccountIndex = u32;
-
-/// Balance of an account.
-pub type Balance = u128;
-
-/// Asset ID for generic asset module.
-pub type AssetId = u32;
-
-/// The runtime supported proof of delegation format.
-pub type Doughnut = DoughnutV0;
-
-/// Type used for expressing timestamp.
-pub type Moment = u64;
-
-/// Index of a transaction in the chain.
-pub type Index = u64;
-
-/// A hash of some data used by the chain.
-pub type Hash = primitives::H256;
-
-/// A timestamp: milliseconds since the unix epoch.
-/// `u64` is enough to represent a duration of half a billion years, when the
-/// time scale is milliseconds.
-pub type Timestamp = u64;
-
-/// Digest item type.
-pub type DigestItem = generic::DigestItem<Hash>;
-/// Header type.
-pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
-/// Block type.
-pub type Block = generic::Block<Header, OpaqueExtrinsic>;
-/// Block ID.
-pub type BlockId = generic::BlockId<Block>;
+pub mod types;
